+++ lib/oakdex/pokedex/invalid_pokemon_error.rb
@@ -1,0 +1,7 @@
+module Oakdex
+  module Pokedex
+    # Exception if custom pokemon is invalid
+    class InvalidPokemonError < StandardError
+    end
+  end
+end