+++ CHANGELOG.md
@@ -2,12 +2,19 @@
 All notable changes to this project will be documented in this file.
 Please keep to the changelog format described on [keepachangelog.com](http://keepachangelog.com).
 
+## [Unreleased]
+
+### Added
+- Added `Oakdex::Pokedex::Pokemon.import!([json1, json2])` to ruby gem in order to add custom pokemon.
+- Added `importPokemon([json1, json2])` to npm package
+
+
 ## [0.3.0] - 2019-01-06
 
 ### Changed
 - Changed signature of all javascript functions: instead of `findPokemon('Eevee', fn)` it is now just `findPokemon('Eevee')` as it is now sync instead of async. More in #81
 
-## Removed
+### Removed
 - Removed Ruby 2.0.0 Support
 
 ### Fixed